+++ src/components/attachments/AttachmentStatus.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Tooltip } from '@material-ui/core'
+import { Tooltip } from '@mui/material'
 import UploadingAttachment from './UploadingAttachment'
 import { FormTypes } from '@oneblink/types'
 import { checkIsUsingLegacyStorage } from '../../services/attachments'